[eudsl-llvmpy] Strict submodule package layout; move the DSL to llvm.dsl
Add thin Python shims (ir.py/types.py/passmanager.py/jit.py) re-exporting each
C-extension submodule, and reconcile intrinsics.py over llvm.eudslllvm_ext.intrinsics.
__init__ no longer star-exports to the package top (strict, MLIR-style): callers
use llvm.ir.Context, llvm.jit.LLJIT, etc. The DSL frontend is now llvm.dsl.function
/ llvm.dsl.yield_ / llvm.dsl.range_. DSL/support modules import from the .ir/.types
C submodules.
