+++ projects/eudsl-llvmpy/src/llvm/__init__.py
@@ -3,12 +3,12 @@
 #  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 #  Copyright (c) 2025.
 
-from .eudslllvm_ext import *
-from .eudslllvm_ext import __doc__
+# MLIR-style submodule layout: nothing is re-exported at the package top level.
+# Use llvm.ir.Context, llvm.types.i32, llvm.passmanager.run_passes,
+# llvm.jit.LLJIT, llvm.intrinsics.sqrt, llvm.dsl.function, etc.
+from . import ir, types, passmanager, jit, intrinsics, dsl  # noqa: F401
 
 from .dsl.values import install_value_casters as _install_value_casters
 
 _install_value_casters()
 
-from .dsl.func import function
-from .dsl.cf import yield_, range_
