+++ projects/eudsl-llvmpy/src/llvm/jit.py
@@ -1,6 +1,4 @@
 #  Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 #  See https://llvm.org/LICENSE.txt for license information.
 #  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-from .func import function  # noqa: F401
-from .cf import yield_, range_  # noqa: F401
-
+from .eudslllvm_ext.jit import *  # noqa: F401,F403
